Module Fmlib_parse.Interfaces

Module types

module type MINIMAL_PARSER = sig ... end

A minimal parser is a sink of tokens which either succeeds or returns a list of failed syntax expectations.

module type NORMAL_PARSER = sig ... end

A normal parser parses a stream of tokens like a MINIMAL_PARSER. In addition it can have a state and semantic errors.

module type FULL_PARSER = sig ... end

A full parser parses a stream of tokens like a MINIMAL_PARSER. In addition it can have a state, semantic errors and gives access to the lookahead tokens.

module type LEXER = sig ... end

A lexer is a restartable parser where the tokens are characters.

module type COMBINATOR = sig ... end

module type CHAR_DECODER = sig ... end

Decoder for unicode characters.

module type CHAR_ENCODER = sig ... end

Encoder for unicode characters.

module type CHAR_CODEC = sig ... end

Encoder and decoder for unicode characters.