Module type Interfaces.CHAR_CODEC

Encoder and decoder for unicode characters.

module Encoder : CHAR_ENCODER

module Decoder : CHAR_DECODER

A character decoder has some valid initial value init. In order to decode a character bytes has to be pushed into the decoder by put until the decoder is either complete (is_complete) or has encoutered some decoding error (has_error).