Module Fmlib_parse.Lexeme

Helper module to make parsing with Character more convenient.

The module needs a subset of the combinators of the module Character and some definitions of whitespace and identifiers in order to provide more convenient combinators than the module Character has.

module Make : sig ... end