Module Fmlib_parse.Utf16

Encoders and Decoders for Unicode Characters encoded in UTF-16.

module Be : sig ... end

Encoder and Decoder for Unicode Characters encoded in UTF-16 Big Endian.

module Le : sig ... end

Encoder and Decoder for Unicode Characters encoded in UTF-16 Little Endian.